refactor(python): modularize Python bridge and expose all ECS/system/event/worldgen APIs
